+++ lib/exmeralda_web/live/admin/library_live/show.ex
@@ -1,6 +1,7 @@
 defmodule ExmeraldaWeb.Admin.LibraryLive.Show do
   use ExmeraldaWeb, :live_view
   import ExmeraldaWeb.Admin.Helper
+  import ExmeraldaWeb.Shared.Helper
   alias Exmeralda.Topics
 
   @impl true
@@ -106,7 +107,7 @@ defmodule ExmeraldaWeb.Admin.LibraryLive.Show do
       >
         <:col :let={ingestion} label="ID" field={:id}>{ingestion.id}</:col>
         <:col :let={ingestion} label="State" field={:state}>
-          <.ingestion_state state={ingestion.state} />
+          <.ingestion_state_badge state={ingestion.state} />
         </:col>
         <:col :let={ingestion} label="Created At" field={:inserted_at}>
           {datetime(ingestion.inserted_at)}
